# internal/api/handlers_sysops.go
package api

import (
	"net/http"
	"net/url"

	"github.com/go-chi/chi/v5"
	"github.com/nodelistdb/internal/database"
	"github.com/nodelistdb/internal/storage"
)

// SysopsHandler handles requests for listing sysops.
// GET /api/sysops?name=John&limit=50&offset=0
func (s *Server) SysopsHandler(w http.ResponseWriter, r *http.Request) {
	// Parse query parameters
	query := r.URL.Query()
	nameFilter := query.Get("name")

	// Parse pagination with custom limits for sysops
	limit, offset := parsePaginationParams(query, 50, 200)

	// Get unique sysops
	sysops, err := s.storage.GetUniqueSysops(r.Context(), nameFilter, limit, offset)
	if err != nil {
		writeStorageErrorf(w, "Failed to get sysops", err)
		return
	}

	if sysops == nil {
		sysops = []storage.SysopInfo{}
	}
	response := map[string]interface{}{
		"sysops": sysops,
		"count":  len(sysops),
		"filter": map[string]interface{}{
			"name":   nameFilter,
			"limit":  limit,
			"offset": offset,
		},
	}

	WriteJSONSuccess(w, response)
}

// SysopNodesHandler handles requests for getting nodes by sysop.
// GET /api/sysops/{name}/nodes?limit=100
func (s *Server) SysopNodesHandler(w http.ResponseWriter, r *http.Request) {
	// Extract sysop name from path using Chi
	sysopName := chi.URLParam(r, "name")
	if sysopName == "" {
		WriteJSONError(w, "Sysop name cannot be empty", http.StatusBadRequest)
		return
	}

	// URL decode the sysop name
	decodedName, err := url.PathUnescape(sysopName)
	if err != nil {
		WriteJSONError(w, "Invalid sysop name encoding", http.StatusBadRequest)
		return
	}

	// Parse limit with higher max for sysop nodes
	limit, _ := parsePaginationParams(r.URL.Query(), 100, 1000)

	// Get nodes for this sysop
	nodes, err := s.storage.GetNodesBySysop(r.Context(), decodedName, limit)
	if err != nil {
		writeStorageErrorf(w, "Failed to get nodes", err)
		return
	}

	if nodes == nil {
		nodes = []database.Node{}
	}
	response := map[string]interface{}{
		"sysop_name": decodedName,
		"nodes":      nodes,
		"count":      len(nodes),
		"limit":      limit,
	}

	WriteJSONSuccess(w, response)
}